The turd passes all tests
